requirements.txt
albumentations==1.1.0
grad_cam==1.3.5
matplotlib==3.3.4
numpy==1.19.2
opencv_python==4.5.4.60
Pillow==9.0.1
scikit_image==0.17.2
torch==1.9.1
torchvision==0.10.1
tqdm==4.62.3
